Add the correct animation control sequence
