+++ docs/Card.html
@@ -1705,7 +1705,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="game_CardManager.js.html">game/CardManager.js</a>, <a href="game_CardManager.js.html#line52">line 52</a>
+        <a href="game_CardManager.js.html">game/CardManager.js</a>, <a href="game_CardManager.js.html#line43">line 43</a>
     </li></ul></dd>
     
 
@@ -2031,7 +2031,7 @@
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 4.0.4</a> on Sun Jun 08 2025 05:53:55 GMT+0000 (Coordinated Universal Time)
+    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 4.0.4</a> on Sun Jun 08 2025 10:08:14 GMT-0700 (北美太平洋夏令时间)
 </footer>
 
 <script> prettyPrint(); </script>
