+++ docs/Enemy.html
@@ -30,8 +30,7 @@
     
         <h2><span class="attribs"><span class="type-signature"></span></span>Enemy<span class="signature">(options)</span><span class="type-signature"></span></h2>
         
-            <div class="class-description">The enemy of game
-An enemy has a name, HP, attack power, and abilities</div>
+            <div class="class-description">The enemy of gameAn enemy has a name, HP, attack power, and abilities</div>
         
     
 </header>
@@ -54,9 +53,7 @@ An enemy has a name, HP, attack power, and abilities</div>
 
 
 <div class="description">
-    Creates a new Enemy instance.*
-This constructor initializes the enemy's name, current HP, maximum HP, and image.
-It also sets up the Shadow DOM to visually display the enemy's status.
+    Creates a new Enemy instance.*This constructor initializes the enemy's name, current HP, maximum HP, and image.It also sets up the Shadow DOM to visually display the enemy's status.
 </div>
 
 
@@ -345,7 +342,7 @@ It also sets up the Shadow DOM to visually display the enemy's status.
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="components_Enemy.js.html">components/Enemy.js</a>, <a href="components_Enemy.js.html#line67">line 67</a>
+        <a href="components_Enemy.js.html">components/Enemy.js</a>, <a href="components_Enemy.js.html#line96">line 96</a>
     </li></ul></dd>
     
 
@@ -417,7 +414,7 @@ It also sets up the Shadow DOM to visually display the enemy's status.
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="components_Enemy.js.html">components/Enemy.js</a>, <a href="components_Enemy.js.html#line20">line 20</a>
+        <a href="components_Enemy.js.html">components/Enemy.js</a>, <a href="components_Enemy.js.html#line23">line 23</a>
     </li></ul></dd>
     
 
@@ -546,7 +543,7 @@ It also sets up the Shadow DOM to visually display the enemy's status.
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="components_Enemy.js.html">components/Enemy.js</a>, <a href="components_Enemy.js.html#line138">line 138</a>
+        <a href="components_Enemy.js.html">components/Enemy.js</a>, <a href="components_Enemy.js.html#line168">line 168</a>
     </li></ul></dd>
     
 
@@ -706,7 +703,7 @@ It also sets up the Shadow DOM to visually display the enemy's status.
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="components_Enemy.js.html">components/Enemy.js</a>, <a href="components_Enemy.js.html#line128">line 128</a>
+        <a href="components_Enemy.js.html">components/Enemy.js</a>, <a href="components_Enemy.js.html#line158">line 158</a>
     </li></ul></dd>
     
 
@@ -794,7 +791,7 @@ It also sets up the Shadow DOM to visually display the enemy's status.
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="components_Enemy.js.html">components/Enemy.js</a>, <a href="components_Enemy.js.html#line117">line 117</a>
+        <a href="components_Enemy.js.html">components/Enemy.js</a>, <a href="components_Enemy.js.html#line147">line 147</a>
     </li></ul></dd>
     
 
@@ -882,7 +879,7 @@ It also sets up the Shadow DOM to visually display the enemy's status.
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="components_Enemy.js.html">components/Enemy.js</a>, <a href="components_Enemy.js.html#line146">line 146</a>
+        <a href="components_Enemy.js.html">components/Enemy.js</a>, <a href="components_Enemy.js.html#line176">line 176</a>
     </li></ul></dd>
     
 
@@ -1019,7 +1016,7 @@ It also sets up the Shadow DOM to visually display the enemy's status.
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="components_Enemy.js.html">components/Enemy.js</a>, <a href="components_Enemy.js.html#line105">line 105</a>
+        <a href="components_Enemy.js.html">components/Enemy.js</a>, <a href="components_Enemy.js.html#line135">line 135</a>
     </li></ul></dd>
     
 
@@ -1329,7 +1326,7 @@ It also sets up the Shadow DOM to visually display the enemy's status.
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="components_Enemy.js.html">components/Enemy.js</a>, <a href="components_Enemy.js.html#line67">line 67</a>
+        <a href="components_Enemy.js.html">components/Enemy.js</a>, <a href="components_Enemy.js.html#line96">line 96</a>
     </li></ul></dd>
     
 
@@ -1401,7 +1398,7 @@ It also sets up the Shadow DOM to visually display the enemy's status.
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="components_Enemy.js.html">components/Enemy.js</a>, <a href="components_Enemy.js.html#line20">line 20</a>
+        <a href="components_Enemy.js.html">components/Enemy.js</a>, <a href="components_Enemy.js.html#line23">line 23</a>
     </li></ul></dd>
     
 
@@ -1530,7 +1527,7 @@ It also sets up the Shadow DOM to visually display the enemy's status.
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="components_Enemy.js.html">components/Enemy.js</a>, <a href="components_Enemy.js.html#line138">line 138</a>
+        <a href="components_Enemy.js.html">components/Enemy.js</a>, <a href="components_Enemy.js.html#line168">line 168</a>
     </li></ul></dd>
     
 
@@ -1690,7 +1687,7 @@ It also sets up the Shadow DOM to visually display the enemy's status.
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="components_Enemy.js.html">components/Enemy.js</a>, <a href="components_Enemy.js.html#line128">line 128</a>
+        <a href="components_Enemy.js.html">components/Enemy.js</a>, <a href="components_Enemy.js.html#line158">line 158</a>
     </li></ul></dd>
     
 
@@ -1778,7 +1775,7 @@ It also sets up the Shadow DOM to visually display the enemy's status.
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="components_Enemy.js.html">components/Enemy.js</a>, <a href="components_Enemy.js.html#line117">line 117</a>
+        <a href="components_Enemy.js.html">components/Enemy.js</a>, <a href="components_Enemy.js.html#line147">line 147</a>
     </li></ul></dd>
     
 
@@ -1866,7 +1863,7 @@ It also sets up the Shadow DOM to visually display the enemy's status.
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="components_Enemy.js.html">components/Enemy.js</a>, <a href="components_Enemy.js.html#line146">line 146</a>
+        <a href="components_Enemy.js.html">components/Enemy.js</a>, <a href="components_Enemy.js.html#line176">line 176</a>
     </li></ul></dd>
     
 
@@ -2003,7 +2000,7 @@ It also sets up the Shadow DOM to visually display the enemy's status.
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="components_Enemy.js.html">components/Enemy.js</a>, <a href="components_Enemy.js.html#line105">line 105</a>
+        <a href="components_Enemy.js.html">components/Enemy.js</a>, <a href="components_Enemy.js.html#line135">line 135</a>
     </li></ul></dd>
     
 
@@ -2055,7 +2052,7 @@ It also sets up the Shadow DOM to visually display the enemy's status.
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 4.0.4</a> on Sun Jun 08 2025 05:53:55 GMT+0000 (Coordinated Universal Time)
+    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 4.0.4</a> on Sun Jun 08 2025 10:08:14 GMT-0700 (北美太平洋夏令时间)
 </footer>
 
 <script> prettyPrint(); </script>
