+++ docs/Player.html
@@ -1145,9 +1145,7 @@
 
 
 <div class="description">
-    Attempts to spend a specified amount of energy.
-If the player does not have enough energy, the operation fails and returns false.
-Otherwise, deducts the energy and returns true.
+    Attempts to spend a specified amount of energy.If the player does not have enough energy, the operation fails and returns false.Otherwise, deducts the energy and returns true.
 </div>
 
 
@@ -1473,7 +1471,7 @@ Otherwise, deducts the energy and returns true.
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 4.0.4</a> on Sun Jun 08 2025 05:53:55 GMT+0000 (Coordinated Universal Time)
+    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 4.0.4</a> on Sun Jun 08 2025 10:08:14 GMT-0700 (北美太平洋夏令时间)
 </footer>
 
 <script> prettyPrint(); </script>
