+++ docs/Powell_Powell.html
@@ -49,9 +49,7 @@
 
 
 <div class="description">
-    Powell constructor
-Uses Enemy cosntructor to initialize enemy
-Assigns couple of private variables for internal use
+    Powell constructorUses Enemy cosntructor to initialize enemyAssigns couple of private variables for internal use
 </div>
 
 
@@ -212,7 +210,7 @@ Assigns couple of private variables for internal use
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 4.0.4</a> on Sun Jun 08 2025 05:53:55 GMT+0000 (Coordinated Universal Time)
+    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 4.0.4</a> on Sun Jun 08 2025 10:08:14 GMT-0700 (北美太平洋夏令时间)
 </footer>
 
 <script> prettyPrint(); </script>
