+++ docs/game_Battle.js.html
@@ -76,8 +76,8 @@ export class Battle {
     document.querySelector('.player .character-container').append(this.player);
     document.querySelector('.enemies .character-container').append(this.enemy);
     document.querySelector('.hand-area').append(this.player.hand);
-    console.log('showing game start banner');
     this.eventBus.publish('startTurn');
+
 
     const discardButton = document.querySelector('.discard-button');
     const discardModal = document.querySelector('.discard-modal');
@@ -97,6 +97,7 @@ export class Battle {
       closeDiscardModal,
       discardPile: this.player.discard
     });
+
     const deckModalInstance = new DeckModal({
       deckButton,
       deckModal,
@@ -131,8 +132,8 @@ export class Battle {
   /**
    * Enemy AI turn logic, then ends turn
    */
-  enemyAction() {
-    this.enemy.takeTurn(this.enemy.HP, this.enemy.maxHP, this.player.HP, this.player.maxHP);
+  async enemyAction() {
+    await this.enemy.takeTurn(this.enemy.HP, this.enemy.maxHP, this.player.state.HP, this.player.state.maxHP);
     // check whether player is dead
     this.eventBus.publish('checkGameOver');
     this.eventBus.publish('endTurn');
@@ -172,24 +173,15 @@ export class Battle {
 
     if (this.battleOver) return;
 
-        // Show banner for each turn
+    // update effect
+    // this.eventBus.publish('onTurnStart', { actor: this.currentActor });
+
     if (this.currentActor === 'player') {
       this.showBanner('Player turn', this.turnCount);
       this.waitForPlayerAction();
     } else {
       this.showBanner('Enemy turn', this.turnCount);
       this.enemyAction();
-    }
-
-    // update effect
-    // this.eventBus.publish('onTurnStart', { actor: this.currentActor });
-
-    if (this.currentActor === 'player') {
-      this.waitForPlayerAction();
-    } else {
-      //wait for 2 seconds before enemy action
-      setTimeout(() => {
-      this.enemyAction(); }, 2000);
     }
   }
 
@@ -209,7 +201,11 @@ export class Battle {
     // }
   }
 
-    /**
+  /**
+   * Clean up all event listeners (optional)
+   */
+
+      /**Add commentMore actions
    * Show a pop-up banner with a message and optional turn number
    * @param {string} message
    * @param {number} [turn]
@@ -240,14 +236,12 @@ showBanner(message, turn) {
   }, 1200);
 }
 
-  /**
-   * Clean up all event listeners (optional)
-   */
   destroy() {
     this.eventBus.unsubscribe('startTurn', this.startTurn);
     this.eventBus.unsubscribe('endTurn', this.handleTurnEnd);
     this.eventBus.unsubscribe('cardPlayed', this.player.playCard);
   }
+
 
 }</code></pre>
         </article>
@@ -265,7 +259,7 @@ showBanner(message, turn) {
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 4.0.4</a> on Sun Jun 08 2025 05:53:55 GMT+0000 (Coordinated Universal Time)
+    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 4.0.4</a> on Sun Jun 08 2025 10:08:14 GMT-0700 (北美太平洋夏令时间)
 </footer>
 
 <script> prettyPrint(); </script>
