+++ docs/game_CardManager.js.html
@@ -27,7 +27,7 @@
     <section>
         <article>
             <pre class="prettyprint source linenums"><code>import {Pile} from '../components/Pile.js';
-import Hand from '../components/Hand.js';
+
 
 /**
  * @deprecated Use Deck.shuffle() instead
@@ -59,15 +59,6 @@ export function moveCard(card, fromCollection, toCollection) {
   // Moving cards between deck, cards, discard, etc.
   //Note: Use the reference of the card when using this function and not it's
   //actual value as it would cause issue with card duplicates.
-  if(fromCollection instanceof Hand){
-    const index = fromCollection.cards.indexOf(card);
-
-    if (index >= 0 &amp;&amp; index &lt; fromCollection.cards.length) {
-      fromCollection.cards.splice(index, 1);
-      toCollection.addCard(card);
-    }
-
-  }
 
   const index = fromCollection.cards.indexOf(card);
 
@@ -99,8 +90,7 @@ export function filterCards(pile, condition) {
   }
 
   return result;  
-}
-</code></pre>
+}</code></pre>
         </article>
     </section>
 
@@ -116,7 +106,7 @@ export function filterCards(pile, condition) {
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 4.0.4</a> on Sun Jun 08 2025 05:53:55 GMT+0000 (Coordinated Universal Time)
+    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 4.0.4</a> on Sun Jun 08 2025 10:08:14 GMT-0700 (北美太平洋夏令时间)
 </footer>
 
 <script> prettyPrint(); </script>
