+++ docs/global.html
@@ -280,8 +280,7 @@
 
 
 <div class="description">
-    Called at the end of the player's turn.
-Discards the player's cards.
+    Called at the end of the player's turn.Discards the player's cards.
 </div>
 
 
@@ -885,8 +884,7 @@ Discards the player's cards.
 
 
 <div class="description">
-    Checks if the player is dead (0 HP).
-If the player does not exist, returns null.
+    Checks if the player is dead (0 HP).If the player does not exist, returns null.
 </div>
 
 
@@ -2219,8 +2217,7 @@ If the player does not exist, returns null.
 
 
 <div class="description">
-    Called at the start of the player's turn.
-Resets energy and draws 5 cards.
+    Called at the start of the player's turn.Resets energy and draws 5 cards.
 </div>
 
 
@@ -2400,7 +2397,7 @@ Resets energy and draws 5 cards.
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 4.0.4</a> on Sun Jun 08 2025 05:53:55 GMT+0000 (Coordinated Universal Time)
+    Documentation generated by <a href="https://github.com/jsdoc/jsdoc">JSDoc 4.0.4</a> on Sun Jun 08 2025 10:08:14 GMT-0700 (北美太平洋夏令时间)
 </footer>
 
 <script> prettyPrint(); </script>
